+++ python/sglang/srt/model_executor/model_runner.py
@@ -32,10 +32,7 @@ from sglang.srt.configs.model_config import (
 )
 from sglang.srt.configs.update_config import adjust_config_with_unaligned_cpu_tp
 from sglang.srt.debug_utils.dumper import dumper
-from sglang.srt.distributed import (
-    bootstrap,
-    parallel_state,
-)
+from sglang.srt.distributed import bootstrap
 from sglang.srt.distributed.device_communicators.mooncake_transfer_engine import (
     maybe_init_shared_mooncake_transfer_engine,
 )
@@ -45,7 +42,8 @@ from sglang.srt.elastic_ep.elastic_ep import (
     ElasticEPStateManager,
     get_healthy_expert_location_src_rank,
     join_process_groups,
-    try_recover_ranks,
+    maybe_rebalance_after_rank_fault,
+    maybe_recover_ep_ranks,
 )
 from sglang.srt.elastic_ep.expert_backup_client import ExpertBackupClient
 from sglang.srt.environ import envs
@@ -178,7 +176,6 @@ from sglang.srt.state_capturer.routed_experts import (
     set_global_experts_capturer,
 )
 from sglang.srt.utils import (
-    broadcast_pyobj,
     cpu_has_amx_support,
     enable_show_time_cost,
     get_available_gpu_memory,
@@ -1488,67 +1485,3 @@ class ModelRunner:
             load_format=load_format,
         )
         self.load_config = load_config
-
-
-def maybe_recover_ep_ranks(
-    *,
-    tp_group: parallel_state.GroupCoordinator,
-    eplb_manager: EPLBManager,
-    random_seed: int,
-) -> bool:
-    # TODO(perf): `active_ranks.all()` on a CUDA tensor triggers host-device
-    # synchronization, and this function is on the forward-path.
-    # This check only runs when `--elastic-ep-backend` is enabled, so the
-    # synchronization overhead does not propagate to other configs.
-    # Leave for future optimization of the elastic EP path.
-    if tp_group.active_ranks.all() and tp_group.active_ranks_cpu.all():
-        return False
-
-    tp_active_ranks = tp_group.active_ranks.detach().cpu().numpy()
-    tp_active_ranks_cpu = tp_group.active_ranks_cpu.detach().numpy()
-    tp_active_ranks &= tp_active_ranks_cpu
-    # NOTE: `ranks_to_recover` uses indices in `tp_group`. For the current
-    # Mooncake elastic EP implementation we assume `--pp-size=1`, so the
-    # tp-group index is the same as the global rank index.
-    ranks_to_recover = [
-        i for i in range(len(tp_active_ranks)) if not tp_active_ranks[i]
-    ]
-
-    # try_recover_ranks polls peer state via Mooncake EP backend.
-    # Mooncake's internal semantics guarantee that all ranks observe
-    # consistent peer readiness state, so collective operations below
-    # are safe even though polling appears local.
-    if ranks_to_recover and try_recover_ranks(ranks_to_recover):
-        eplb_manager.reset_generator()
-        broadcast_global_expert_location_metadata(
-            src_rank=get_healthy_expert_location_src_rank(
-                invoked_in_elastic_ep_rejoin_path=False
-            )
-        )
-        ElasticEPStateManager.instance().reset()
-        broadcast_pyobj(
-            [random_seed],
-            parallel_state.get_world_group().rank,
-            parallel_state.get_world_group().cpu_group,
-            src=parallel_state.get_world_group().ranks[0],
-        )
-        logger.info(f"recover ranks {ranks_to_recover} done")
-        return True
-
-    return False
-
-
-def maybe_rebalance_after_rank_fault(*, eplb_manager: EPLBManager) -> bool:
-    elastic_ep_state = ElasticEPStateManager.instance()
-    if elastic_ep_state is None or elastic_ep_state.is_active_equal_last():
-        return False
-    elastic_ep_state.snapshot_active_to_last()
-    elastic_ep_state.sync_active_to_cpu()
-    logger.info("EPLB due to rank faults")
-    gen = eplb_manager.rebalance()
-    while True:
-        try:
-            next(gen)
-        except StopIteration:
-            break
-    return True
